requirements.txt
# Framework web
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9

# Pipeline OCR
opencv-python-headless>=4.8.0
imutils>=0.5.4
scikit-image>=0.21.0
pytesseract>=0.3.10
easyocr>=1.7.0
python-doctr[torch]>=0.8.0

# IA / extraction
groq>=0.9.0

# Persistance
requests>=2.31.0

# Utilitaires
python-dotenv>=1.0.0
Pillow>=10.0.0
numpy>=1.24.0
